Item_Background.pixil - See desc.
From pixilart.com

Here's a background asset for all tools and items (for example weapons) in games for practicing. The colour codes below can be used to remove the background for when the items/assets are added into the games. It creates a more vibrant and easier-to-see backdrop, especially if parts of the asset will be white.

Colour codes: #cfd8dc, #9e9e9e
Clickup: CU-22wgxh

@IrisDroidology

#1 + Learn art