Idle animations based on player input for movement direction
Having trouble with transitioning between idle blend tree and walk blend tree for the animator. The boolean `isMoving` that we set can't have a value changed - what I mean by this is that we want the blend tree to be `Walk` when `isMoving` parameter is `true`, but in the Animator inspector we can't set a value for this. Maybe it's because the variable name is already defined in `Scripts/PlayerController.cs` 
It's something we need to figure out

https://github.com/acord-robotics/unity-intro/issues/21